# openapi.yaml
openapi: 3.0.3
info:
  title: pubky-http-relay
  description: |
    An HTTP relay for asynchronous message passing with store-and-forward semantics.

    **Two endpoint groups are available:**
    - `/inbox/{id}` - **Recommended.** Store-and-forward with explicit ACK support
    - `/link/{id}` - **Deprecated.** Synchronous producer/consumer pairing

    ## Inbox API (recommended)

    The inbox API provides at-least-once delivery with explicit acknowledgment:

    1. **Producer** stores a message with `POST /inbox/{id}`
    2. **Consumer** retrieves it with `GET /inbox/{id}`
    3. **Consumer** acknowledges with `DELETE /inbox/{id}`
    4. **Producer** waits for ACK with `GET /inbox/{id}/await` (or polls `/ack`)

    Messages persist for 5 minutes (configurable) until acknowledged. The consumer
    can retry `GET` after connection drops and still receive the message. The
    producer knows delivery succeeded only after the consumer explicitly ACKs.

    This pattern is mobile-friendly: if the app is backgrounded or killed, the
    message remains available for retry, and the producer can detect timeouts.
  version: 0.6.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /link/{id}:
    get:
      summary: Consume a message (deprecated)
      deprecated: true
      description: |
        **Deprecated.** Use `/inbox` endpoints for new integrations.

        Consumer retrieves a message from the channel. Blocks until a producer
        sends data or the timeout (10 minutes) is reached.
      operationId: linkGet
      tags:
        - link
      parameters:
        - $ref: '#/components/parameters/channelId'
      responses:
        '200':
          $ref: '#/components/responses/MessageReceived'
        '408':
          $ref: '#/components/responses/Timeout'

    post:
      summary: Produce a message (deprecated)
      deprecated: true
      description: |
        **Deprecated.** Use `/inbox` endpoints for new integrations.

        Producer sends a message to the channel. Blocks until a consumer
        retrieves the data or the timeout (10 minutes) is reached.
      operationId: linkPost
      tags:
        - link
      parameters:
        - $ref: '#/components/parameters/channelId'
      requestBody:
        $ref: '#/components/requestBodies/Message'
      responses:
        '200':
          $ref: '#/components/responses/MessageDelivered'
        '408':
          $ref: '#/components/responses/Timeout'
components:
  parameters:
...
    channelId:
      name: id
      in: path
      required: true
      description: |
        Unique channel identifier. Producers and consumers using the same
        ID are paired.
      schema:
        type: string
      example: my-channel
  requestBodies:
    Message:
      description: |
        Message payload (max 2KB). Content-Type is preserved and returned
        to the consumer.
      required: true
      content:
        application/octet-stream:
          schema:
            type: string
            format: binary
            maxLength: 2048
        application/json:
          schema:
            type: object
          example:
            hello: world
        text/plain:
          schema:
            type: string
          example: Hello, world!
  responses:
...
    MessageReceived:
      description: Message successfully retrieved from producer
      headers:
        Content-Type:
          description: Content-Type from the producer's request (if provided)
          schema:
            type: string
      content:
        application/octet-stream:
          schema:
            type: string
            format: binary
        application/json:
          schema:
            type: object
        text/plain:
          schema:
            type: string
    MessageDelivered:
      description: Message successfully delivered to consumer
    Timeout:
      description: Timeout waiting for counterpart (producer or consumer)
...